DR13 2D Histogram (for Region of Interest)
This is the latest 2D Histogram for DR13 region of interest where potential binaries may reside. All of these stars lie on the far right region of the 2D histogram for all of DR13. These stars have satisfied the requirements that log(max x-range) greater than or equal to 1.83 and log(R min) is less than or equal to 1.08 (the values provided are typical ranges that binaries tend to fall in according to analyzing our known binaries). The known binaries examined came from DR12, more specifically Drew's List and SB9 SB2s.